refactor(security): migrate to EndeavourSecurity service and remove Casdoor/Casbin dependencies
Replaced Casdoor and Casbin authentication/authorization implementation with new EndeavourSecurity service that delegates to a dedicated auth server. Updated all API endpoints and middleware to use the centralized security service instead of direct Casdoor/Casbin calls. Removed unused AWS SDK and related packages. Cleaned up configuration and type definitions accordingly.

This change migrates the entire authentication and authorization logic to use the new EndeavourSecurity service, which handles all security operations by delegating to a dedicated authentication server. The migration includes updating API endpoints, middleware, and removing all references to Casdoor and Casbin components while maintaining the same security functionality.

Closes: rich/endeavour-security
