--- public/casbin/model.conf
@@ -1,15 +1,0 @@
-# Request definition
-[request_definition]
-r = sub, obj, act
-
-# Policy definition
-[policy_definition]
-p = sub_rule, obj, act, eft
-
-# Matchers
-[matchers]
-m = eval(p.sub_rule) && keyMatch(r.obj, p.obj) && keyMatch(r.act, p.act)
-
-# Policy effect
-[policy_effect]
-e = some(where (p.eft == allow))
