--- public/casbin/policy.csv
@@ -1,15 +1,0 @@
-p, (r.sub != null) && (r.sub.admin == true), *, *, allow
-
-p, true, /callback*, *, allow
-p, true, /socket.io/*, *, allow
-p, true, /api/auth/*, *, allow
-p, true, /api/oauth/*, *, allow
-p, true, /api/public/*, *, allow
-p, (r.sub != null), /api/*, *, allow
-p, true, /__nuxt_error*, *, allow
-
-p, true, /, *, allow
-p, true, /login, *, allow
-p, true, /callback, *, allow
-p, true, /unauthorized, *, allow
-p, (r.sub != null) && ('Endeavour/Admin' in (r.sub.groups)), /QueryRunner, *, allow
