+++ server/api/auth/logout.ts
@@ -1,4 +1,4 @@
 
 export default defineEventHandler(async (event) => {
-  await globalThis.authenticator.logout(event);
+  await globalThis.security.logout(event);
 });
