+++ server/api/query/run.post.ts
@@ -51,7 +51,7 @@ defineRouteMeta({
 });
 
 export default defineEventHandler(async (event) => {
-  const currentUser = await globalThis.authenticator.getUser(event);
+  const currentUser = await globalThis.security.getUser(event);
 
   const userId = currentUser!.id;
   const userName = currentUser!.userName;
