--- server/utils/security/auth/auth.base.ts
@@ -1,56 +1,0 @@
-import type {EventHandlerRequest, H3Event } from "h3";
-import { setCookie, getCookie, deleteCookie } from "h3";
-import type {User} from "~~/models/User";
-
-export default abstract class Authenticator {
-  protected readonly ACCESS_TOKEN = "access_token";
-  protected readonly USER_SESSION = "user_session";
-  protected readonly REFRESH_TOKEN = "refresh_token";
-
-  abstract getLoginUrl(redirectUri: string, state: string): Promise<string>;
-  abstract getTokensFromCodeInternal(code: string, redirectUri: string): Promise<{ accessToken: string, refreshToken?: string}>;
-  abstract getUserInternal(accessToken: string): Promise<User>;
-  abstract revokeTokens(accessToken?: string, refreshToken?: string): Promise<void>;
-  abstract introspect(accessToken?: string, refreshToken?: string): Promise<void>;
-
-  public async getTokensFromCode(event: H3Event<EventHandlerRequest>, code: string, redirectUri: string) {
-    const { accessToken, refreshToken } = await this.getTokensFromCodeInternal(code, redirectUri);
-    this.setCookies(event, accessToken, refreshToken)
-  }
-
-  public setCookies(event: H3Event<EventHandlerRequest>, accessToken: string, refreshToken?: string) {
-    setCookie(event, this.ACCESS_TOKEN, accessToken);
-    setCookie(event, this.USER_SESSION, JSON.stringify(this.getUserInternal(accessToken)));
-
-    if (refreshToken) {
-      setCookie(event, this.REFRESH_TOKEN, refreshToken);
-    } else {
-      deleteCookie(event, this.REFRESH_TOKEN);
-    }
-  }
-
-  public async getUser(event: H3Event<EventHandlerRequest>) : Promise<User | undefined> {
-    const accessToken = getCookie(event, this.ACCESS_TOKEN);
-    if (!accessToken)
-      return undefined;
-
-    return await this.getUserInternal(accessToken);
-  }
-
-  public async logout(event: H3Event<EventHandlerRequest>) {
-    const accessToken = getCookie(event, this.ACCESS_TOKEN);
-    const refreshToken = getCookie(event, this.REFRESH_TOKEN);
-
-    await this.revokeTokens(accessToken, refreshToken);
-
-    deleteCookie(event, this.ACCESS_TOKEN);
-    deleteCookie(event, this.REFRESH_TOKEN);
-    deleteCookie(event, this.USER_SESSION);
-  }
-
-  public async requireUser(event: H3Event<EventHandlerRequest>) {
-    const accessToken = getCookie(event, this.ACCESS_TOKEN);
-    const refreshToken = getCookie(event, this.REFRESH_TOKEN);
-    await this.introspect(accessToken, refreshToken);
-  }
-}
