--- server/utils/security/guard/guard.base.ts
@@ -1,7 +1,0 @@
-export default interface Guard {
-  hasPermission(
-    accessToken: string,
-    object: string,
-    action: string
-  ): Promise<boolean>;
-}
