+++ shared/global.d.ts
@@ -1,14 +1,11 @@
 import { Server as IOServer } from "socket.io";
-import type Guard from "~~/server/utils/security/guard/guard.base";
-import type Authenticator from "~~/server/utils/security/auth/auth.base";
+import EndeavourSecurity from "~~/server/utils/endeavour.security";
 
 declare global {
   var io: IOServer;
-  var authenticator: Authenticator;
-  var guard: Guard;
+  var security: EndeavourSecurity;
   interface GlobalThis {
     io: IOServer;
-    authenticator: Authenticator;
-    guard: Guard;
+    security: EndeavourSecurity;
   }
 }
